archive/requirements_ascii.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0
xlrd>=2.0.1
pdfplumber>=0.11.0
pymupdf>=1.24.0
scikit-learn>=1.5.0
xgboost>=2.0.0
lightgbm>=4.3.0
imbalanced-learn>=0.12.0
joblib>=1.3.0
shap>=0.45.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0